+++ services/rfq/relayer/service/statushandler.go
@@ -9,6 +9,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/jellydator/ttlcache/v3"
+	"github.com/synapsecns/sanguine/core/mapmutex"
 	"github.com/synapsecns/sanguine/core/metrics"
 	"github.com/synapsecns/sanguine/services/rfq/api/client"
 	"github.com/synapsecns/sanguine/services/rfq/relayer/chain"
@@ -49,6 +50,8 @@ type QuoteRequestHandler struct {
 	// mutexMiddlewareFunc is used to wrap the handler in a mutex middleware.
 	// this should only be done if Handling, not forwarding.
 	mutexMiddlewareFunc func(func(ctx context.Context, span trace.Span, req reldb.QuoteRequest) error) func(ctx context.Context, span trace.Span, req reldb.QuoteRequest) error
+	// handlerMtx is the mutex for relaying.
+	handlerMtx mapmutex.StringMapMutex
 }
 
 // Handler is the handler for a quote request.
@@ -77,6 +80,7 @@ func (r *Relayer) requestToHandler(ctx context.Context, req reldb.QuoteRequest) 
 		claimCache:          r.claimCache,
 		apiClient:           r.apiClient,
 		mutexMiddlewareFunc: r.mutexMiddleware,
+		handlerMtx:          r.handlerMtx,
 	}
 
 	// wrap in deadline middleware since the relay has not yet happened
@@ -99,12 +103,29 @@ func (r *Relayer) requestToHandler(ctx context.Context, req reldb.QuoteRequest) 
 
 func (r *Relayer) mutexMiddleware(next func(ctx context.Context, span trace.Span, req reldb.QuoteRequest) error) func(ctx context.Context, span trace.Span, req reldb.QuoteRequest) error {
 	return func(ctx context.Context, span trace.Span, req reldb.QuoteRequest) (err error) {
-		unlocker, ok := r.relayMtx.TryLock(hexutil.Encode(req.TransactionID[:]))
+		unlocker, ok := r.handlerMtx.TryLock(hexutil.Encode(req.TransactionID[:]))
 		if !ok {
-			span.SetAttributes(attribute.Bool("locked", true))
+			span.SetAttributes(
+				attribute.Bool("locked", true),
+				attribute.StringSlice("current_locks", r.handlerMtx.Keys()),
+			)
 			return nil
 		}
 		defer unlocker.Unlock()
+
+		// make sure the status has not changed since we last saw it
+		dbReq, err := r.db.GetQuoteRequestByID(ctx, req.TransactionID)
+		if err != nil {
+			return fmt.Errorf("could not get request: %w", err)
+		}
+		if dbReq.Status != req.Status {
+			span.SetAttributes(
+				attribute.Bool("status_changed", true),
+				attribute.String("db_status", dbReq.Status.String()),
+				attribute.String("handler_status", req.Status.String()),
+			)
+			return nil
+		}
 
 		return next(ctx, span, req)
 	}
@@ -121,7 +142,7 @@ func (r *Relayer) deadlineMiddleware(next func(ctx context.Context, span trace.S
 
 		// if deadline < now, we don't even have to bother calling the underlying function
 		if req.Transaction.Deadline.Cmp(big.NewInt(almostNow.Unix())) < 0 {
-			err := r.db.UpdateQuoteRequestStatus(ctx, req.TransactionID, reldb.DeadlineExceeded)
+			err := r.db.UpdateQuoteRequestStatus(ctx, req.TransactionID, reldb.DeadlineExceeded, &req.Status)
 			if err != nil {
 				return fmt.Errorf("could not update request status: %w", err)
 			}
@@ -233,14 +254,19 @@ func (q *QuoteRequestHandler) Handle(ctx context.Context, request reldb.QuoteReq
 // Forward forwards a quote request.
 // this ignores the mutex middleware.
 func (q *QuoteRequestHandler) Forward(ctx context.Context, request reldb.QuoteRequest) (err error) {
+	txID := hexutil.Encode(request.TransactionID[:])
 	ctx, span := q.metrics.Tracer().Start(ctx, fmt.Sprintf("forward-%s", request.Status.String()), trace.WithAttributes(
-		attribute.String("transaction_id", hexutil.Encode(request.TransactionID[:])),
+		attribute.String("transaction_id", txID),
 	))
 	defer func() {
 		metrics.EndSpanWithErr(span, err)
 	}()
 
-	// TODO: consider adding a lock attempt/fail here as a defensive coding strategy. We *expect* stuff to be locked by the time we get to forward.
+	// sanity check to make sure that the lock is already acquired for this tx
+	_, ok := q.handlerMtx.TryLock(txID)
+	if ok {
+		panic(fmt.Sprintf("attempted forward while lock was not acquired for tx: %s", txID))
+	}
 
 	return q.handlers[request.Status](ctx, span, request)
 }